Unnest MoodleAssignmentFeedback
